Add ./configure option "--with-varpath=PATH"
On Unix-like platforms, it is often desirable to install files that
will be modified after installation under a separate directory tree
from that containing read-only data files. This is so that read-only
files can reside in a read-only file system, and also package managers
can update or remove read-only files without worrying about removing
user data.

The new option "--with-varpath=PATH", when specified, configures the
game so that certain directories, namely apex, bone, data, edit, user,
and save, will be created under the given path, as opposed to the path
specified with "--with-libpath=PATH". When the option is not given,
those directories will be created under the lib directory as before.